renamed ecr to cim and removed unwanted files
