+++ examples/cim/dqn/components/dist_actor.py
@@ -1,25 +1,23 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-import os
 import io
 import yaml
 
 import numpy as np
 
 from maro.simulator import Env
-from maro.rl import SimpleLearner, SimpleActor, AgentMode, KStepExperienceShaper, TwoPhaseLinearExplorer
-from maro.utils import Logger, convert_dottable
-from examples.ecr.dqn.components.state_shaper import ECRStateShaper
-from examples.ecr.dqn.components.action_shaper import ECRActionShaper
-from examples.ecr.dqn.components.experience_shaper import TruncatedExperienceShaper
-from examples.ecr.dqn.components.agent_manager import DQNAgentManager
+from maro.rl import AgentMode, SimpleActor, ActorWorker, KStepExperienceShaper, TwoPhaseLinearExplorer
+from maro.utils import convert_dottable
+from examples.cim.dqn.components.state_shaper import ECRStateShaper
+from examples.cim.dqn.components.action_shaper import ECRActionShaper
+from examples.cim.dqn.components.experience_shaper import TruncatedExperienceShaper
+from examples.cim.dqn.components.agent_manager import DQNAgentManager
 
 
 with io.open("config.yml", "r") as in_file:
     raw_config = yaml.safe_load(in_file)
     config = convert_dottable(raw_config)
-
 
 if __name__ == "__main__":
     env = Env(config.env.scenario, config.env.topology, durations=config.env.durations)
@@ -36,17 +34,17 @@ if __name__ == "__main__":
                           "with_cache": config.exploration.with_cache
                           }
     explorer = TwoPhaseLinearExplorer(agent_id_list, config.general.total_training_episodes, **exploration_config)
-    agent_manager = DQNAgentManager(name="ecr_learner",
-                                    mode=AgentMode.TRAIN_INFERENCE,
+    agent_manager = DQNAgentManager(name="cim_remote_actor",
                                     agent_id_list=agent_id_list,
+                                    mode=AgentMode.INFERENCE,
                                     state_shaper=state_shaper,
                                     action_shaper=action_shaper,
                                     experience_shaper=experience_shaper,
                                     explorer=explorer)
-    learner = SimpleLearner(trainable_agents=agent_manager,
-                            actor=SimpleActor(env=env, inference_agents=agent_manager),
-                            logger=Logger("single_host_ecr_learner", auto_timestamp=False))
-
-    learner.train(total_episodes=config.general.total_training_episodes)
-    learner.test()
-    learner.dump_models(os.path.join(os.getcwd(), "models"))
+    proxy_params = {"group_name": config.distributed.group_name,
+                    "expected_peers": config.distributed.actor.peer,
+                    "redis_address": (config.distributed.redis.host_name, config.distributed.redis.port)
+                    }
+    actor_worker = ActorWorker(local_actor=SimpleActor(env=env, inference_agents=agent_manager),
+                               proxy_params=proxy_params)
+    actor_worker.launch()
