--- examples/ecr/distributed_mode/start_redis.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-# start Redis server
-nohup redis-server &
-echo "Redis server up and running"
